requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
confluent-kafka==2.3.0
redis==5.0.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pytest==7.2.0